Promote Caveats up into MethodologyNote + add transparency / report-it line
Brian: '[the methodology note] maybe replaces caveats (and notes section
at the bottom entirely) and makes it front and center clear at the top
before people start using the tool' + 'Indicate that we try our best to
be as transparent as possible about where a number is coming from, but
we may be missing some. If something seems off, please report it!'

Changes:
- Move the 'Models always simplify…' Caveats paragraph from Notes.tsx
  (bottom of page) into MethodologyNote (top of page, above Customize).
  The honesty pass now lands first, where it shapes how the user
  reads everything that follows.
- Add a transparency + report-it line in MethodologyNote linking to
  the existing source-report issue template
  (.github/ISSUE_TEMPLATE/source-report.yml — same audit:report
  channel community feedback already uses).
- Reduce Notes.tsx to just the PageSources footer (sources-backing-this-
  calculation grounding stays — citation convention requires it on
  every data-driven page).
